// src/types/view-data/page-type.types.ts
// Page/Views rendered by popup
export type PageType =
  | 'PRODUCT_PAGE'
  | 'CART_PAGE'
  | 'BOOKMARKS_PAGE'
  | 'SEARCH_RESULTS_PAGE'
  | 'UNSUPPORTED_PAGE'
  | 'UNKNOWN_RETAILER_PAGE'
  | 'UNSUPPORTED_DOMAIN'
  | 'CART_IGNORE_PAGE'
  | 'PRODUCT_IGNORE_PAGE'
  | 'ERROR_PAGE'
  | 'URL_ERROR_PAGE'
  | 'RETAILER_ERROR_PAGE'
  | 'RATE_LIMITED_PAGE'
  | 'SERVICE_UNAVAILABLE_PAGE'
  | 'NON_HTTP_PAGE'
  | 'ERROR'
  | null